+++ healthhive-backend/src/main/java/com/healthhive/model/HealthData.java
@@ -21,7 +21,7 @@ public class HealthData {
     private int steps;
     private int calories;
     private double weight;
-    private int sleepHours;
+    private double sleepHours;
     private String mood;
     private LocalDateTime date;
 }
